+++ src/language/stella-module.ts
@@ -19,6 +19,7 @@ import { SemanticTokenProvider } from "./lsp/semantic-tokens.js";
 import { CompletionProvider } from "./lsp/completion-provider.js";
 import { StellaScopeProvider } from "./references/scope-provider.js";
 import { StellaScopeComputation } from "./references/scope-computation.js";
+import { HoverProvider } from "./lsp/hover-provider.js";
 
 /**
  * Declaration of custom services - add your own service classes here.
@@ -50,6 +51,7 @@ export const StellaModule: Module<
   lsp: {
     SemanticTokenProvider: (services) => new SemanticTokenProvider(services),
     CompletionProvider: (services) => new CompletionProvider(services),
+    HoverProvider: (services) => new HoverProvider(services),
   },
   references: {
     ScopeProvider: (services) => new StellaScopeProvider(services),