+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Validation for unique function names and for the presence of a `main` function.
 - Add copy code/output buttons to the playground.
+- Hover provider with link to documentation for extensions.
 
 ### Changed